+++ DSL/Resql/ljvis/POST/user_group/list_user_groups.sql
@@ -60,8 +60,8 @@ SELECT
 FROM latest l
 WHERE
     (
-        :organisation_id = 0
-        OR :organisation_id = ANY(l.organisations)
+        :organisation_id::BIGINT = 0
+        OR :organisation_id::BIGINT = ANY(l.organisations)
     )
     AND (
         COALESCE(:search, '') = ''
